Refactored event component
